+++ src/downloader.ts
@@ -4,13 +4,13 @@ import path from 'path';
 import { actions, fs, log, selectors, types, util } from 'vortex-api';
 import axios from 'axios';
 
-import { GAME_ID } from './common';
+import { GAME_ID, NOTIF_ID_REQUIREMENTS } from './common';
 import { IPluginRequirement, IGithubDownload } from './types';
 
 export async function download(api: types.IExtensionApi, requirements: IPluginRequirement[], force?: boolean) {
   api.sendNotification({
-    id: 'ue4ss-download-notification',
-    message: 'Installing UE4SS',
+    id: NOTIF_ID_REQUIREMENTS,
+    message: 'Installing Palworld Requirements',
     type: 'activity',
     noDismiss: true,
     allowSuppress: false,
@@ -24,11 +24,17 @@ export async function download(api: types.IExtensionApi, requirements: IPluginRe
       if (force !== true && mod?.id !== undefined) {
         batchActions.push(actions.setModEnabled(profileId, mod.id, true));
         batchActions.push(actions.setModAttribute(GAME_ID, mod.id, 'customFileName', req.userFacingName));
+        batchActions.push(actions.setModAttribute(GAME_ID, mod.id, 'description', 'This is a Palworld modding requirement - leave it enabled.'));
         continue;
       }
       if (req?.modId !== undefined) {
         await downloadNexus(api, req);
       } else {
+        const dlId = req.findDownloadId(api);
+        if (dlId) {
+          await installDownload(api, dlId, req.userFacingName);
+          continue;
+        }
         const asset = await getLatestReleaseDownloadUrl(api, req);
         const tempPath = path.join(util.getVortexPath('temp'), asset.fileName);
         await doDownload(asset.url, tempPath);
@@ -37,41 +43,52 @@ export async function download(api: types.IExtensionApi, requirements: IPluginRe
     }
   } catch (err) {
     // Fallback here.
-    log('error', 'failed to download ue4ss', err);
+    log('error', 'failed to download requirements', err);
     return;
   } finally {
     if (batchActions.length > 0) {
       util.batchDispatch(api.store, batchActions);
     }
-    api.dismissNotification('ue4ss-download-notification');
+    api.dismissNotification(NOTIF_ID_REQUIREMENTS);
   }
+}
+
+async function installDownload(api: types.IExtensionApi, dlId: string, name: string): Promise<void> {
+  return new Promise<void>((resolve, reject) => {
+    api.events.emit('start-install-download', dlId, true, (err, modId) => {
+      if (err !== null) {
+        api.showErrorNotification('Failed to install requirement', err, { allowReport: false });
+        return reject(err);
+      }
+
+      const state = api.getState();
+      const profileId = selectors.lastActiveProfileForGame(state, GAME_ID);
+      const batch = [
+        actions.setModAttributes(GAME_ID, modId, {
+          installTime: new Date(),
+          name,
+        }),
+        actions.setModEnabled(profileId, modId, true),
+      ];
+      util.batchDispatch(api.store, batch);
+      return resolve();
+    })
+  })
 }
 
 async function importAndInstall(api: types.IExtensionApi, filePath: string, name: string) {
   return new Promise<void>((resolve, reject) => {
-    api.events.emit('import-downloads', [filePath], (dlIds: string[]) => {
+    api.events.emit('import-downloads', [filePath], async (dlIds: string[]) => {
       const id = dlIds[0];
       if (id === undefined) {
         return reject(new util.NotFound(filePath));
       }
-      api.events.emit('start-install-download', id, true, (err, modId) => {
-        if (err !== null) {
-          api.showErrorNotification('Failed to install repo', err, { allowReport: false });
-          return reject(err);
-        }
-    
-        const state = api.getState();
-        const profileId = selectors.lastActiveProfileForGame(state, GAME_ID);
-        const batch = [
-          actions.setModAttributes(GAME_ID, modId, {
-            installTime: new Date(),
-            name,
-          }),
-          actions.setModEnabled(profileId, modId, true),
-        ];
-        util.batchDispatch(api.store, batch);
+      try {
+        await installDownload(api, id, name);
         return resolve();
-      });
+      } catch (err) {
+        return reject(err);
+      }
     });
   })
 }