+++ src/modTypes.ts
@@ -1,12 +1,32 @@
 /* eslint-disable */
 import path from 'path';
-import { selectors, types } from 'vortex-api';
-import { BPPAK_MODSFOLDER_PATH, PAK_MODSFOLDER_PATH,
-  PAK_EXTENSIONS, IGNORE_CONFLICTS, LUA_EXTENSIONS } from './common';
-import { resolveUE4SSPath } from './util';
+import { log, selectors, types } from 'vortex-api';
+import {
+  BPPAK_MODSFOLDER_PATH, PAK_MODSFOLDER_PATH,
+  PAK_EXTENSIONS, IGNORE_CONFLICTS, LUA_EXTENSIONS, UE_PAK_TOOL_FILES, MOD_TYPE_PAK, MOD_TYPE_BP_PAK
+} from './common';
+import { resolveUE4SSPath, findInstallFolderByFile } from './util';
+
+import { listPak } from './unrealPakParser';
 
 //#region Utility
 const hasModTypeInstruction = (instructions: types.IInstruction[]) => instructions.find(instr => instr.type === 'setmodtype');
+
+const runPakTool = async (api: types.IExtensionApi, instructions: types.IInstruction[]) => {
+  let modDir: string = undefined;
+  const filtered = instructions
+    .filter((inst: types.IInstruction) => (inst.type === 'copy') && (path.extname(inst.source) === '.pak'));
+  for (const pak of filtered) {
+    if (!modDir) {
+      modDir = await findInstallFolderByFile(api, pak.source);
+      if (!modDir) {
+        return null;
+      }
+    }
+    const data = await listPak(api, path.join(modDir, pak.source));
+    return data.modType;
+  }
+}
 //#endregion
 
 //#region MOD_TYPE_PAK
@@ -19,15 +39,23 @@ export function getPakPath(api: types.IExtensionApi, game: types.IGame) {
   return pakPath;
 }
 
-export function testPakPath(instructions: types.IInstruction[]): Promise<boolean> {
-  // Pretty basic set up right now.
+export async function testPakPath(api: types.IExtensionApi, instructions: types.IInstruction[]): Promise<boolean> {
   if (hasModTypeInstruction(instructions)) {
     return Promise.resolve(false);
+  }
+  try {
+    const modType = await runPakTool(api, instructions);
+    if (modType && modType === MOD_TYPE_PAK) {
+      return true;
+    }
+  } catch (err) {
+    // Pak tool fudged up - resume default stop pattern installation.
+    log('error', 'Failed to ascertain modType using pak tool', err);
   }
   const filteredPaks = instructions
     .filter((inst: types.IInstruction) => (inst.type === 'copy')
       && (PAK_EXTENSIONS.includes(path.extname(inst.source as any))));
- 
+
   const excludeInstructions: types.IInstruction[] = instructions.filter((inst => {
     if (inst.type !== 'copy') return false;
     const segments = inst.source.split(path.sep);
@@ -61,7 +89,7 @@ export function testLUAPath(instructions: types.IInstruction[]): Promise<boolean
   const filtered = instructions
     .filter((inst: types.IInstruction) => (inst.type === 'copy')
       && (LUA_EXTENSIONS.includes(path.extname(inst.source as any))));
- 
+
   const supported = filtered.length > 0;
   return Promise.resolve(supported) as any;
 }
@@ -77,9 +105,17 @@ export function getBPPakPath(api: types.IExtensionApi, game: types.IGame) {
   return luaPath;
 }
 
-export function testBPPakPath(instructions: types.IInstruction[]): Promise<boolean> {
+export async function testBPPakPath(api: types.IExtensionApi, instructions: types.IInstruction[]): Promise<boolean> {
   if (hasModTypeInstruction(instructions)) {
     return Promise.resolve(false);
+  }
+  try {
+    const modType = await runPakTool(api, instructions);
+    if (modType && modType === MOD_TYPE_BP_PAK) {
+      return true;
+    }
+  } catch (err) {
+    // Pak tool fudged up - resume default stop pattern installation.
   }
   const filteredPaks = instructions
     .filter((inst: types.IInstruction) => {
@@ -95,7 +131,7 @@ export function testBPPakPath(instructions: types.IInstruction[]): Promise<boole
       }
       return true;
     });
- 
+
   const excludeInstructions: types.IInstruction[] = instructions.filter((inst => {
     if (inst.type !== 'copy') return false;
     const segments = inst.source.toLowerCase().split(path.sep);
@@ -106,6 +142,21 @@ export function testBPPakPath(instructions: types.IInstruction[]): Promise<boole
   }))
 
   const supported = filteredPaks.length > 0 && excludeInstructions.length === 0;
+  return Promise.resolve(supported) as any;
+}
+//#endregion
+
+//#region UnrealPakTool
+// Pak tool only needs a test function.
+export function testUnrealPakTool(instructions: types.IInstruction[]): Promise<boolean> {
+  if (hasModTypeInstruction(instructions)) {
+    return Promise.resolve(false);
+  }
+  const filtered = instructions
+    .filter((inst: types.IInstruction) => (inst.type === 'copy')
+      && UE_PAK_TOOL_FILES.includes(path.basename(inst.source)));
+
+  const supported = filtered.length === UE_PAK_TOOL_FILES.length;
   return Promise.resolve(supported) as any;
 }
 //#endregion